#1038 [MODIFY] 출석체크 상태에 따라 버튼 메시지 분기
